refactor(frontend): rename entity-provenance.ts to entity-sources.ts
Match the companion component name (EntitySources.svelte). Remove the
corresponding follow-up from SourcesFollowUps.md.
